requirement.txt
Flask==3.0.3
MySQL-connector python==9.0.0